Time: 0 ms (100%), Space: 10.4 MB (97.64%) - LeetHub

+++ 0410-split-array-largest-sum/0410-split-array-largest-sum.cpp
@@ -44,6 +44,6 @@ public:
             }
         }
 
-        return (double)left;
+        return (int)left;
     }
 };
